+++ src/async_thing.ts
@@ -14,6 +14,18 @@ function isAsyncIterable<T>(iterable: AnyIterable<T>): iterable is AsyncIterable
 }
 
 export class AsyncThing<T> extends AsyncThingMath<T> {
+    public static empty<T>() {
+        return new AsyncThing<T>((() => ({
+            next: () => Promise.resolve({ done: true }),
+        })));
+    }
+
+    public static single<T>(item: T) {
+        return new AsyncThing<T>(async function* single() {
+            yield item;
+        });
+    }
+
     public readonly isAsyncThing = true;
 }