+++ src/index.ts
@@ -1,3 +1,4 @@
 import { thing } from "./async_thing";
 
+export { AsyncThing } from "./async_thing";
 export default thing;